+++ client/src/main/pages/SumOfRanks/SumOfRanksPage.module.css
@@ -1,0 +1,9 @@
+.ranksTable {
+  overflow-x: auto;
+  font-size: 11px;
+}
+
+.ranksTable td{
+  padding: 10px;
+  text-align: center;
+}
