+++ server/src/main/java/org/worldcubeassociation/statistics/dto/rank/SumOfRanksDto.java
@@ -7,8 +7,6 @@ import java.util.List;
 @Data
 public class SumOfRanksDto {
     private Integer regionRank;
-    private String region;
-    private String regionType;
     private String wcaId;
     private Integer overall;
     private List<SumOfRankEventDto> events;
